Create algolia components
