+++ packages/web/components/Algolia/index.js
@@ -1,0 +1,11 @@
+export { default as AlgoliaSearchProvider } from './provider';
+export { default as SearchBox } from './searchBox';
+export { default as Stats } from './stats';
+export { default as SortBy } from './sortBy';
+export { default as HitsPerPage } from './hitsPerPage';
+export { default as HitCard } from './hitCard';
+export { default as Pagination } from './pagination';
+export { default as ClearRefinements } from './clearRefinements';
+export { default as Panel } from './panel';
+export { default as RefinementList } from './refinementList';
+export { default as ToggleRefinement } from './toggleRefinement';
