--- packages/web/providers/index.js
@@ -1,2 +1,0 @@
-export { default as AlgoliaSearchProvider } from './AlgoliaSearchProvider';
-export { default as ThemeProvider } from './ThemeProvider';
