+++ packages/web/stories/Card.stories.js
@@ -30,7 +30,7 @@ export const Grid = () => {
 						title,
 						category,
 						price,
-						image,
+						logo,
 						place,
 						date,
 						likes,
@@ -43,7 +43,7 @@ export const Grid = () => {
 							title={title}
 							category={category}
 							price={price}
-							image={image}
+							logo={logo}
 							place={place}
 							date={date}
 							likes={likes}
