+++ packages/web/styles/ThemeProvider.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { ThemeProvider as StyledThemeProvider } from 'styled-components';
-import { theme } from '../styles';
+import theme from './theme';
 
 const ThemeProvider = ({ children }) => (
 	<StyledThemeProvider theme={theme}>{children}</StyledThemeProvider>
