+++ packages/web/styles/index.js
@@ -1,2 +1,3 @@
-export { default as global } from './global';
+export { default as GlobalStyle } from './global';
 export { default as theme } from './theme';
+export { default as ThemeProvider } from './ThemeProvider';
